pyproject.toml
[project]
name = "hops-v1"
version = "0.1.0"
description = "HOPS_V1 Etsy Bulk Product"
authors = [
    {name = "Cüneyt Vural",email = "asrinwallprintweb@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
packages = [{include = "core"}]
requires-python = ">=3.11,<3.14"
dependencies = [
    "pyside6 (>=6.9.2,<7.0.0)",
    "pywin32 (>=311,<312)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^6.0.1"
pyinstaller = "^6.16.0"

